+++ TktkDirectX12GameLib/TktkDX12GameLib/inc/TktkDX12BaseComponents/3D/MeshDrawer/SphereMeshDrawerUseResourceHandles.h
@@ -5,10 +5,10 @@ namespace tktk
 {
 	struct SphereMeshDrawerUseResourceHandles
 	{
-		unsigned int rtvDescriptorHeapHandle;
-		unsigned int cameraHandle;
-		unsigned int shadowMapCameraHandle;
-		unsigned int lightHandle;
+		size_t rtvDescriptorHeapHandle;
+		size_t cameraHandle;
+		size_t shadowMapCameraHandle;
+		size_t lightHandle;
 	};
 }
 #endif // !SPHERE_MESH_DRAWER_USE_RESOURCE_HANDLES_H_
