+++ TktkDirectX12GameLib/TktkDX12GameLib/inc/TktkDX12Game/DXGameResource/Billboard/BillboardDrawFuncBaseArgs.h
@@ -6,16 +6,16 @@ namespace tktk
 	struct BillboardDrawFuncBaseArgs
 	{
 		// 使用するビューポートのハンドル
-		unsigned int			viewportHandle;
+		size_t	viewportHandle;
 
 		// 使用するシザー矩形のハンドル
-		unsigned int			scissorRectHandle;
+		size_t	scissorRectHandle;
 
 		// 使用するレンダーターゲット用のディスクリプタヒープハンドル
-		unsigned int			rtvDescriptorHeapHandle;
+		size_t	rtvDescriptorHeapHandle;
 
 		// 使用する深度ステンシル用のディスクリプタヒープハンドル
-		unsigned int			dsvDescriptorHeapHandle;
+		size_t	dsvDescriptorHeapHandle;
 	};
 }
 #endif // !BILLBOARD_DRAW_FUNC_BASE_ARGS_H_
