+++ TktkDirectX12GameLib/TktkDX12GameLib/inc/TktkDX12Game/DXGameResource/Billboard/BillboardMaterial.h
@@ -18,16 +18,16 @@ namespace tktk
 	public:
 
 		// 「BillboardMaterialData」のインスタンスを作り、そのリソースのハンドルを返す
-		unsigned int create(const BillboardMaterialInitParam& initParam);
+		size_t create(const BillboardMaterialInitParam& initParam);
 
 		// 指定したビルボードを描画する
-		void drawBillboard(unsigned int handle, const BillboardDrawFuncBaseArgs& drawFuncArgs) const;
+		void drawBillboard(size_t handle, const BillboardDrawFuncBaseArgs& drawFuncArgs) const;
 
 		// 引数が表すコピーバッファを使って座標変換情報を管理する定数バッファを更新する
-		void updateTransformCbuffer(unsigned int handle, unsigned int copyBufferHandle, const BillboardCbufferUpdateFuncArgs& updateArgs) const;
+		void updateTransformCbuffer(size_t handle, size_t copyBufferHandle, const BillboardCbufferUpdateFuncArgs& updateArgs) const;
 
 		// 引数が表すコピーバッファを使って座標変換情報を管理する定数バッファを更新する（切り抜き範囲指定版）
-		void updateTransformCbufferUseClippingParam(unsigned int handle, unsigned int copyBufferHandle, const BillboardCbufferUpdateFuncArgs& updateArgs, const BillboardClippingParam& clippingParam) const;
+		void updateTransformCbufferUseClippingParam(size_t handle, size_t copyBufferHandle, const BillboardCbufferUpdateFuncArgs& updateArgs, const BillboardClippingParam& clippingParam) const;
 
 	private:
